+++ app/page.tsx
@@ -1,5 +1,27 @@
 import "./globals.css";
-import SignIn from "@/pages/signIn";
-export default function Home() {
-  return <SignIn />;
-}
+import React from "react";
+import "./page.css";
+import Box from "@mui/material/Box";
+import WbSunny from "@mui/icons-material/WbSunny";
+import NightsStayIcon from "@mui/icons-material/NightsStay";
+import { yellow } from "@mui/material/colors";
+import SigninForm from "@/components/signinForm";
+
+const SignIn: React.FC = () => {
+  return (
+    <div className="signin">
+      <div className="signin-icon">
+        <WbSunny fontSize="large" sx={{ color: yellow[500] }} />
+        <NightsStayIcon fontSize="large" color="secondary" />
+      </div>
+      <h1 className="signIn-heading">Login to your Account</h1>
+      <div className="signInForm-wrapper">
+        <Box className="signIn-form">
+          <SigninForm />
+        </Box>
+      </div>
+    </div>
+  );
+};
+
+export default SignIn;
