+++ components/current-weather/currentWeather.css
@@ -1,7 +1,6 @@
 .currentWeather-container {
   color: white;
   min-height: 20.6rem;
-  max-width: 29.3rem;
   border-radius: 1rem;
   padding: 1rem;
   border: 2px solid yellow;
