+++ components/navbar/navbar.css
@@ -4,8 +4,10 @@
   justify-content: space-between;
   background-color: rgba(0, 0, 0, 0.459);
   color: white;
-  padding: 0.5rem;
+  padding: 0.5rem 1rem;
 }
 .heading {
   font-size: clamp(1rem, 1rem + 0.5vw, 2rem);
+  color: rgb(212, 212, 33);
+  font-weight: 700;
 }
