+++ components/navbar/navbar.tsx
@@ -4,14 +4,17 @@ import "./navbar.css";
 import profilepic from "./images/alex-suprun-ZHvM3XIOHoE-unsplash.jpg";
 import LogoutForm from "../logoutForm";
 import { auth } from "@/auth";
-const Navbar: React.FC = async () => {
-  const session = await auth();
+import { redirect } from "next/navigation";
+const Navbar: React.FC = () => {
   return (
     <>
-      <div className="heading">WEATHER APP</div>
-      <div className="user-image">
+      <div className="heading">WEATHER APP 🌞</div>
+      <div
+        className="user-image"
+        style={{ display: "flex", alignItems: "center" }}
+      >
         <IconButton>
-          <Avatar src={session?.user?.image} />
+          <Avatar variant="rounded" sx={{ height: "40px", width: "40px" }} />
         </IconButton>
         <LogoutForm />
       </div>
