--- components/sidebar/sidebar.css
@@ -1,26 +1,0 @@
-.sidebar-main {
-  width: 100%;
-  max-width: 80px;
-  height: 100%;
-  border-radius: 15px;
-  background-color: hsla(240, 65%, 3%, 0.549);
-  color: white;
-  border: 2px solid yellow;
-}
-.sidebar-top-container {
-  display: flex;
-  justify-content: center;
-}
-.weather-text {
-  font-size: 10px;
-}
-.sidebar-divider {
-  background-color: white;
-}
-.sidebar-bottom {
-  margin-top: 200px;
-}
-.button.active {
-  border-left: 5px solid yellow;
-  transition: border-left 0.15s ease;
-}
